FAC-963:impl:Move GenerateAllAssessmentInsights use case to insight package
