+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/GenerateAllAssessmentInsightsRestController.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.adapter.in.rest.assessment;
+package org.flickit.assessment.core.adapter.in.rest.insight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.assessment.GenerateAllAssessmentInsightsUseCase;
-import org.flickit.assessment.core.application.port.in.assessment.GenerateAllAssessmentInsightsUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.GenerateAllAssessmentInsightsUseCase;
+import org.flickit.assessment.core.application.port.in.insight.GenerateAllAssessmentInsightsUseCase.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
