+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/GenerateAllAssessmentInsightsServiceTest.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.service.assessment;
+package org.flickit.assessment.core.application.service.insight;
 
 import org.flickit.assessment.common.application.MessageBundle;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
@@ -11,21 +11,21 @@ import org.flickit.assessment.common.exception.CalculateNotValidException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.common.exception.ValidationException;
 import org.flickit.assessment.core.application.domain.*;
-import org.flickit.assessment.core.application.port.in.assessment.GenerateAllAssessmentInsightsUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.GenerateAllAssessmentInsightsUseCase.Param;
 import org.flickit.assessment.core.application.port.out.assessment.GetAssessmentProgressPort;
-import org.flickit.assessment.core.application.port.out.assessmentinsight.CreateAssessmentInsightPort;
-import org.flickit.assessment.core.application.port.out.assessmentinsight.LoadAssessmentInsightPort;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.flickit.assessment.core.application.port.out.attribute.CreateAttributeScoresFilePort;
 import org.flickit.assessment.core.application.port.out.attribute.LoadAttributesPort;
-import org.flickit.assessment.core.application.port.out.attributeinsight.CreateAttributeInsightPort;
-import org.flickit.assessment.core.application.port.out.attributeinsight.LoadAttributeInsightsPort;
 import org.flickit.assessment.core.application.port.out.attributevalue.LoadAttributeValuePort;
+import org.flickit.assessment.core.application.port.out.insight.assessmentinsight.CreateAssessmentInsightPort;
+import org.flickit.assessment.core.application.port.out.insight.assessmentinsight.LoadAssessmentInsightPort;
+import org.flickit.assessment.core.application.port.out.insight.attributeinsight.CreateAttributeInsightPort;
+import org.flickit.assessment.core.application.port.out.insight.attributeinsight.LoadAttributeInsightsPort;
+import org.flickit.assessment.core.application.port.out.insight.subjectinsight.CreateSubjectInsightPort;
+import org.flickit.assessment.core.application.port.out.insight.subjectinsight.LoadSubjectInsightsPort;
 import org.flickit.assessment.core.application.port.out.maturitylevel.LoadMaturityLevelsPort;
 import org.flickit.assessment.core.application.port.out.minio.UploadAttributeScoresFilePort;
 import org.flickit.assessment.core.application.port.out.subject.LoadSubjectsPort;
-import org.flickit.assessment.core.application.port.out.subjectinsight.CreateSubjectInsightPort;
-import org.flickit.assessment.core.application.port.out.subjectinsight.LoadSubjectInsightsPort;
 import org.flickit.assessment.core.application.port.out.subjectvalue.LoadSubjectValuePort;
 import org.flickit.assessment.core.test.fixture.application.*;
 import org.junit.jupiter.api.Test;
